LLM-Orchestrated-Multi-Model-Traffic-Analysis-System-main/backend/requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-filter==23.5

# WebSocket and Real-time Communication
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
websockets==12.0

# Task Queue and Background Processing
celery==5.3.4

# Database Drivers
pymongo==4.6.0
mongoengine==0.27.0
dnspython==2.4.2

# Configuration and Environment
python-decouple==3.8
python-dotenv==1.0.0

# Image Processing and Computer Vision
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning and AI
ultralytics>=8.2.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scikit-learn>=1.4.0

# LLM Integration
openai==1.3.7
anthropic==0.7.8
groq>=0.4.1

# Data Analysis and Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# System Monitoring and Performance
psutil==5.9.6

# File Handling and Utilities
python-multipart==0.0.6
PyYAML>=6.0

# Security and Cryptography
cryptography==41.0.7
bcrypt==4.1.2

# Web Server
gunicorn==21.2.0

# Development and Debugging (optional)
ipython>=8.0.0
jupyter>=1.0.0

# Additional utilities for enhanced functionality
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.65.0

# Object detection and tracking enhancements
filterpy>=1.4.5
imutils>=0.5.4

# Statistical analysis
statsmodels>=0.14.0